refactor(storage): relocate EmbeddingSpace/SpaceStatus DTOs to me-types
`EmbeddingSpace`, `SpaceStatus`, and the `DEFAULT_NAME` constant were pure data
types (plain derives, no rusqlite coupling) living in me-backend-sqlite's
store/embedding_spaces.rs. The upcoming me-backend-postgres carve needs the
same DTOs for its SchemaManager fingerprint methods, so they move to me-types
(L0) — the layer both backend leaves already depend on — rather than creating
a backend-to-backend dependency.

me-backend-sqlite re-exports them from store::embedding_spaces for path
stability, so every existing internal caller and the facade's
inspect/restore.rs keep resolving unchanged. `default_active` moves from
`#[cfg(test)]` to `#[cfg(feature = "test-util")]`, since cross-crate cfg(test)
does not reach the downstream crate's own tests. No tests move — the
accessor-function tests stay in me-backend-sqlite.

Part of Wave 2 (#816) S2, sub-PR 3 (closes the S2 slice #930 tracker).
